python/requirements.txt
river>=0.21
pandas>=2.0
pymysql>=1.1
python-dotenv>=1.0
click>=8.1
